random id bug on vector<Vertex>, graphviewer is working fine

+++ src/GraphViewer.cpp
@@ -93,35 +93,35 @@ bool GraphViewer::closeWindow() {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::addNode(int id) {
+bool GraphViewer::addNode(unsigned long long id) {
 	char buff[200];
 	sprintf(buff, "addNode1 %d\n", id);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::addNode(int id, int x, int y) {
+bool GraphViewer::addNode(unsigned long long id, int x, int y) {
 	char buff[200];
 	sprintf(buff, "addNode3 %d %d %d\n", id, x, y);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::addEdge(int id, int v1, int v2, int edgeType) {
+bool GraphViewer::addEdge(unsigned long long id, unsigned long long v1, unsigned long long v2, int edgeType) {
 	char buff[200];
 	sprintf(buff, "addEdge %d %d %d %d\n", id, v1, v2, edgeType);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeLabel(int k, string label) {
+bool GraphViewer::setEdgeLabel(unsigned long long k, string label) {
 	char buff[200];
 	sprintf(buff, "setEdgeLabel %d %s\n", k, label.c_str());
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setVertexLabel(int k, string label) {
+bool GraphViewer::setVertexLabel(unsigned long long k, string label) {
 	char buff[200];
 	sprintf(buff, "setVertexLabel %d %s\n", k, label.c_str());
 	string str(buff);
@@ -135,28 +135,28 @@ bool GraphViewer::defineEdgeColor(string color) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::removeNode(int id) {
+bool GraphViewer::removeNode(unsigned long long id) {
 	char buff[200];
 	sprintf(buff, "removeNode %d\n", id);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::removeEdge(int id) {
+bool GraphViewer::removeEdge(unsigned long long id) {
 	char buff[200];
 	sprintf(buff, "removeEdge %d\n", id);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeColor(int k, string color) {
+bool GraphViewer::setEdgeColor(unsigned long long k, string color) {
 	char buff[200];
 	sprintf(buff, "setEdgeColor %d %s\n", k, color.c_str());
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeThickness(int k, int thickness) {
+bool GraphViewer::setEdgeThickness(unsigned long long k, int thickness) {
 	char buff[200];
 	sprintf(buff, "setEdgeThickness %d %d\n", k, thickness);
 	string str(buff);
@@ -170,7 +170,7 @@ bool GraphViewer::defineVertexColor(string color) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setVertexColor(int k, string color) {
+bool GraphViewer::setVertexColor(unsigned long long k, string color) {
 	char buff[200];
 	sprintf(buff, "setVertexColor %d %s\n", k, color.c_str());
 	string str(buff);
@@ -184,14 +184,14 @@ bool GraphViewer::setBackground(string path) {
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeWeight(int id, int weight) {
+bool GraphViewer::setEdgeWeight(unsigned long long id, int weight) {
 	char buff[200];
 	sprintf(buff, "setEdgeWeight %d %d\n", id, weight);
 	string str(buff);
 	return con->sendMsg(str);
 }
 
-bool GraphViewer::setEdgeFlow(int id, int flow) {
+bool GraphViewer::setEdgeFlow(unsigned long long id, int flow) {
 	char buff[200];
 	sprintf(buff, "setEdgeFlow %d %d\n", id, flow);
 	string str(buff);
