+++ src/SharingLocation.h
@@ -7,7 +7,7 @@ private:
 	int maxCapacity;
 	int slots;
 public:
-	SharingLocation(int id, double latitude, double longitude, double altitude, 
+	SharingLocation(unsigned long long id, double latitude, double longitude, double altitude,
 		std::string name, int maxCapacity, int slots);
 	bool depositBike(int number = 1);
 	bool liftBike(int number = 1);
